Update code to Python 3 from Python 2 This runs (passes all tests) on Python 3.5.0rc1. Most of the syntactical changes are a result of running the program '2to3'. Some other changes were necessary to make the code Python 3 compatible. For example:  - Python 3 has a clearer distinction between 'bytes' and 'str'.    Unlike in Python 2, in Python 3 (b'abc' != 'abc'). This    required various changes, particularly when using numpy    arrays of chars. For example, np.fromstring('abc', dtype='S1')    gives the numpy array [b'a', b'b', b'c'], which is difficult    for comparison. Changing this to np.fromiter('abc', dtype='U1')    helps because it gives the numpy array ['a', 'b', 'c'] --    i.e., the characters are Unicode rather than bytes. Similarly,    the multiprocessing.sharedctypes.RawArray, with the    ctypes.c_char_p type, stores bytes in Python 3 and, unlike in    Python 2, are not directly comparable to Unicode chars. Thus,    it is necessary to encode()/decode() strings/bytes when working    with these arrays.  - Division in Python 3 no longer floors by default -- e.g.,    3/2==1.5 in Python 3, but 3/2==1 in Python 2. This required    a few changes.

+++ hybseldesign/utils/interval.py
@@ -103,11 +103,13 @@ class IntervalSet(object):
             # interval. Then determine whether curr_endpoint is contained
             # within each of the "current" intervals.
             if self_endpoint_index < 2 * len(self.intervals):
-                curr_self_interval = self.intervals[self_endpoint_index / 2]
+                curr_self_interval = self.intervals[
+                    int(self_endpoint_index / 2)]
             else:
                 curr_self_interval = None
             if other_endpoint_index < 2 * len(other.intervals):
-                curr_other_interval = other.intervals[other_endpoint_index / 2]
+                curr_other_interval = other.intervals[
+                    int(other_endpoint_index / 2)]
             else:
                 curr_other_interval = None
